+++ sem_seo_toolkit/models/sem_geotarget.py
@@ -4,15 +4,23 @@ from odoo import api, fields, models
 
 class SemGeoTarget(models.Model):
 
+    #TODO rename to search context as device user agent and map zoom level have nothing to do with a geograpic location
     _name = "sem.geo_target"
 
+    def _default_device_id(self):
+        return self.env['ir.model.data'].get_object('sem_seo_toolkit', 'sem_device_desktop')
+
     search_engine_id = fields.Many2one('sem.search_engine', string="Search Engine")
-    name = fields.Char(string="Name", help="Can be changed to a more meaningful name for reports")
+    name = fields.Char(string="Name", help="A human meaningful name to describe the geographic location / search context")
     location_id = fields.Char(string="Location ID", help="The ID of the location within the search engine")
-    latitude = fields.Float(string="Latitude")
-    longitude = fields.Float(string="Longitude")
+    latitude = fields.Char(string="Latitude")
+    longitude = fields.Char(string="Longitude")
     accuracy_radius_meters = fields.Float(string="Accuracy Radius (Meters)")
-    device_id = fields.Many2one('sem.geo_target.device', string="Device")
+    device_id = fields.Many2one('sem.geo_target.device', default=_default_device_id, string="Device")
+    map_zoom_level = fields.Integer(default=14, string="Map Zoom Level")
+
+    def get_ranking(self, domain, keyword):
+        return self.search_engine_id.get_ranking(self, domain, keyword)
 
 class SemGeoTargetDevice(models.Model):
 
@@ -39,13 +47,13 @@ class SemGeoTargetWizard(models.TransientModel):
 
             # Now add the records to the pool so they can be selected using the geo_target_ids fields
             for geo_target in geo_targets:
-                self.env['sem.geo_target.wizard.record'].create({'location_id': geo_target[0], 'name': geo_target[1]})
+                self.env['sem.geo_target.wizard.record'].create(geo_target)
 
     def add_geo_locations(self):
 
         for geo_target in self.geo_target_ids:
             # Add the geo target to the local cache so it is easy to select locale for future keywords
-            local_geo_target = self.env['sem.geo_target'].create({'search_engine_id': self.search_engine_id.id, 'location_id': geo_target.location_id, 'name': geo_target.name})
+            local_geo_target = self.env['sem.geo_target'].create({'search_engine_id': self.search_engine_id.id, 'location_id': geo_target.location_id, 'name': geo_target.name, 'latitude': geo_target.latitude, 'longitude': geo_target.longitude, 'accuracy_radius_meters': geo_target.accuracy_radius_meters})
 
             # Also add the geo targets to the keyword
             self.keyword_id.geo_target_ids = [(4, local_geo_target.id)]
@@ -56,3 +64,6 @@ class SemGeoTargetWizardRecord(models.TransientModel):
 
     location_id = fields.Char(string="The reference ID of the location in the Search Engine database")
     name = fields.Char(string="Name")
+    latitude = fields.Char(string="Latitude")
+    longitude = fields.Char(string="Longitude")
+    accuracy_radius_meters = fields.Float(string="Accuracy Radius (Meters)")